+++ sdk/core/iOS/MoEReactBridge/MoEngageReactConstants.m
@@ -43,3 +43,4 @@ NSString* const kInAppSelfHandled   = @"MoEInAppCampaignSelfHandled";
 NSString* const kPushTokenGenerated = @"MoEPushTokenGenerated";
 NSString* const kPushClicked        = @"MoEPushClicked";
 NSString* const kPermissionResult   = @"MoEPermissionResult";
+NSString* const kLogoutComplete     = @"MoELogoutComplete";